backend/requirements.txt
Django==5.1.3
djangorestframework==3.16.1
joblib==1.5.0
openai>=1.0.0
pandas==2.3.3
scikit-learn==1.6.1
